fixed searchError
