+++ db/database.js
@@ -3,7 +3,16 @@
  */
 var config = require('../cfg/config.json');
 var request = require('request');
-var searchError = require('./searchError.js');
+var SearchError = function (message, date, searchedName){
+    this.name='SearchError';
+    this.message= message || 'Some Failure happened while searching for a SentimentAnalysis';
+    this.stack= (new Error()).stack;
+    this.date= date;
+    this.searchedName= searchedName;
+};
+
+SearchError.prototype = Object.create(Error.prototype);
+SearchError.prototype.constructor = SearchError;
 /*
  Saves a json to the character in the database
  json format:
